requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.10.2
google-auth==2.6.6
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-cloud-videointelligence==2.8.3
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio-status==1.48.2
grpcio==1.51.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
Jinja2==3.1.2
jsonschema==4.17.3
keras
MarkupSafe==2.1.1
numpy==1.23.5
opencv-python-headless
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
proto-plus==1.22.1
protobuf==3.20.3
pyarrow==10.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
pytz==2022.6
requests==2.28.1
rich==12.6.0
rsa==4.9
scikit-learn
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit
tensorflow
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.1.9
zipp==3.11.0